requirements.txt
# Newton Supercomputer - Python Dependencies
# Used by Vercel (@vercel/python) and Render for deployment

# Web Framework
fastapi>=0.109.0
uvicorn>=0.27.0
starlette>=0.36.0

# Data Validation
pydantic>=2.5.0

# Cryptography (for Vault AES-256-GCM encryption)
cryptography>=41.0.0

# HTTP Client (for external integrations)
requests>=2.31.0

# External Search (for grounding)
googlesearch-python>=1.2.0

# Graph operations (for TLM / bridge)
networkx>=3.1

# Deployment (Render/traditional)
gunicorn>=21.2.0

# AWS Lambda adapter
mangum>=0.17.0

# Flask (for ProjectADA and legacy endpoints)
flask>=3.0.0

# CAD Designer dependencies
Django>=4.2
djangorestframework>=3.14
shapely>=2.0.0
ezdxf>=1.1.0
cadquery>=2.3.0
